requirements.txt
django
django-allauth
djangorestframework
graphene_django
graphene-file-upload
djangorestframework-jwt
django-graphql-jwt
django-cors-headers
mysqlclient
channels

apollo-client
apollo-boost
react-apollo

requests


